Rename: answer api 구조 수정
