--- src/api/services/answer/common/index.ts
@@ -1,15 +1,0 @@
-import { authorizationInstance } from '@/api/instance'
-
-type AnswerQuestionParam = {
-  friendId: number
-}
-
-export const answerRandomQuestion = async ({
-  friendId,
-}: AnswerQuestionParam) => {
-  const response = await authorizationInstance.post('/api/answer/common', {
-    friendId,
-  })
-
-  return response.data
-}
